+++ sorting/exchange_sorts.h
@@ -4,16 +4,16 @@
 #include <stddef.h>
 
 
-void bubble_sort(int arr[], size_t n);
+void bubble_sort(int arr[], int n);
 
 
 // a variant of bubble sort (double side bubble)
-void cocktail_sort(int arr[], size_t n);
+void cocktail_sort(int arr[], int n);
 
 
 void gneric_quicksort(void *base, size_t num, size_t size,
 				int (*cmp_func)(const void*, const void*));
-void quicksort(int arr[], size_t n);
+void quicksort(int arr[], int n);
 
 
 #endif //_ALGOFUN_EXCHANGE_SORTS_H
